Update README.md
It is very important to determine the original request (specifically, the original `url`, because the `url` given by `$info` variable is not much useful as it may be the result of redirection, as opposed to the `url` given in `request`. So, let people know it. 

( btw, would be nice to mention what is "group" class and it's usage ).